Fixes pre-commit errors and warnings
